remove orderbook subscription hook and other fixes
